+++ Datamigratie.Server/Features/Mapping/GlobalMapping/Models/GlobalConfigurationResponseModel.cs
@@ -1,0 +1,7 @@
+namespace Datamigratie.Server.Features.Mapping.GlobalMapping.Models;
+
+public class GlobalConfigurationResponseModel
+{
+    public string? Rsin { get; set; }
+    public DateTime? UpdatedAt { get; set; }
+}