+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [Migrate all versions of a Document from the e-Suite to OpenZaak](https://dimpact.atlassian.net/browse/DATA-119)
 - [Store e-Suite data as document in OpenZaak](https://dimpact.atlassian.net/browse/DATA-102)
 - [Migration of non-configurable additional Document properties from the e-Suite to OpenZaak)](https://dimpact.atlassian.net/browse/DATA-112)
+- [Global configuration of RSIN](https://dimpact.atlassian.net/browse/DATA-139)
 - [Migration log overview migration-level](https://dimpact.atlassian.net/browse/DATA-17)